move CoordinateInfo to module info
