+++ src/modules/commons/index.js
@@ -8,4 +8,3 @@ import './components/spinner';
 import './components/overflowMenu';
 import './components/lazy';
 import './components/valueSelect';
-import './components/coordinateInfo';
