+++ src/modules/info/components/coordinateInfo/CoordinateInfo.js
