+++ src/modules/info/components/coordinateInfo/index.js
