+++ src/modules/info/i18n/coordinateInfo.provider.js
