+++ src/modules/info/i18n/index.js
@@ -1,0 +1,4 @@
+import { $injector } from '../../../injection';
+import { provide as coordinateInfoProvider } from './coordinateInfo.provider';
+const { TranslationService: translationService } = $injector.inject('TranslationService');
+translationService.register('coordinateInfoProvider', coordinateInfoProvider);
