+++ src/modules/info/index.js
@@ -1,0 +1,2 @@
+import './i18n';
+import './components/coordinateInfo';
